relate item with merchant

+++ app/Models/Order.php
@@ -5,6 +5,7 @@ namespace App\Models;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
+
 class Order extends Model
 {
     use HasFactory;
@@ -21,13 +22,13 @@ class Order extends Model
 
     public function getFeatureDiscounts()
     {
-        return (int)$this->features->reduce(function ($carry, $feature) {
+        return (float)$this->features->reduce(function ($carry, $feature) {
             return $feature->pivot->discount * $feature->pivot->quantity + $carry;
         }, 0);
     }
 
     public function paidAmount()
     {
-        return (int)$this->payments->reduce(fn ($carry, $payment) => $payment->pivot->amount + $carry, 0);
+        return (float)$this->payments->reduce(fn ($carry, $payment) => $payment->pivot->amount + $carry, 0);
     }
 }
